Spread out bis form into multiple components
Planning for something BIG
